requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
pydantic[email]==2.9.2
python-dotenv==1.0.1
groq==0.11.0
requests==2.32.3